+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added `plot_learning_curve` and corresponding `result.plot.learning_curve`
 - Added `plot_validation_curve` and corresponding `result.plot.validation_curve`
 - Replaced `permutation_importance` with scikit-learn's implementation
+- Added `target_correlation` plots to Dataset.plot
 
 
 # v0.9.2